requirements.txt
elevenlabs
python-dotenv
soundfile
fastapi[standard]